benchmarks/longmemeval-mem0/requirements.txt
# Mem0 LongMemEval Benchmark Dependencies
# Tested with Python 3.12+
# Install: pip install -r requirements.txt

mem0ai==1.0.10
openai==2.30.0
nltk==3.9.4
pydantic==2.12.5
